crushed the project into modules
