+++ .gitignore
@@ -10,6 +10,8 @@ Thumbs.db
 # Вспомогательные файлы
 *.log*
 node_modules/
+bundle.js
+bundle.js.map
 
 # Папка с собранными файлами проекта
 build/
